Pull agents test fixtures out to _test_agents.py
