+++ tests/mlmodel_langchain/test_chain.py
@@ -38,6 +38,8 @@ from newrelic.api.llm_custom_attributes import WithLlmCustomAttributes
 from newrelic.api.transaction import add_custom_attribute
 from newrelic.common.object_names import callable_name
 
+from ._test_agents import exercise_agent
+
 try:
     from langchain_classic.chains import create_retrieval_chain
     from langchain_classic.chains.combine_documents import create_stuff_documents_chain
